requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.1
torchvision>=0.16
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0

# ── Market Data ───────────────────────────────────────────────────────────────
yfinance>=0.2.36
requests>=2.31

# ── Technical Indicators ──────────────────────────────────────────────────────
# ta-lib is optional (requires C lib); we ship pure-Python fallbacks
# ta-lib>=0.4.28

# ── Visualisation ─────────────────────────────────────────────────────────────
plotly>=5.18
matplotlib>=3.8
mplfinance>=0.12

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.30

# ── API Server ────────────────────────────────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9

# ── Hyperparameter Optimisation ───────────────────────────────────────────────
optuna>=3.5

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66
Pillow>=10.0

# ── Dev / Test (optional) ─────────────────────────────────────────────────────
# pytest>=7.4
# black>=23.0
# ruff>=0.1
